requirements.txt
numpy==1.23.4
albumentations==1.3.0
boto3==1.26.100
colorama==0.4.6
comet_ml==3.33.3
coremltools==6.3.0
filterpy==1.4.5
Flask==2.2.3
ipython==8.11.0
matplotlib==3.7.1
mss==9.0.1
onnx==1.14.0
onnxruntime==1.15.0
onnxsim==0.4.28
openai==0.27.2
opencv-python-headless
openvino==2022.3.0
paddlepaddle==2.1.3
pafy==0.5.5
pandas==1.5.3
Pillow==9.5.0
psutil==5.9.4
pycocotools==2.0.6
PyYAML==6.0
Requests==2.31.0
scipy==1.10.1
seaborn==0.12.2
setuptools==67.6.0
Shapely==2.0.1
scikit-image
streamlit==1.20.0
streamlit_extras==0.2.7
tensorflow
tensorflow_intel
tensorflowjs
tensorrt==8.6.1
tflite_runtime==2.12.0
tflite_support==0.4.3
thop==0.1.1.post2209072238
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
tritonclient==2.33.0
ultralytics==8.0.99
wget==3.2
x2paddle==1.4.1
